Feature/document contents controller

+++ DocumentDataAPI/DocumentDataAPI/Data/Repositories/ISourceRepository.cs
@@ -4,5 +4,5 @@ namespace DocumentDataAPI.Data.Repositories;
 
 public interface ISourceRepository : IRepository<SourceModel>
 {
-    public int GetCountFromId(int id);
+    public long GetCountFromId(long id);
 }
